int16_t is pointless

+++ src/tools/quemap/lightmap.h
@@ -39,8 +39,8 @@ typedef struct {
 	mat4_t matrix;
 	mat4_t inverse_matrix;
 	vec2_t st_mins, st_maxs;
-	int16_t w, h;
-	int16_t s, t;
+	int32_t w, h;
+	int32_t s, t;
 	luxel_t *luxels;
 	size_t num_luxels;
 	SDL_Surface *ambient;
